Update the FLIP session to use the FlipIOBuffer. Also removed some of the testing hacks for URL rewriting that are no longer needed.
This change removes the 'batching' of frames written
to the socket.  The reason for doing this is because
I'm going to need to start notifying to the upper layer
as progress is made on the writes (e.g. upload notifications).
When the frames are batched (potentially from different 
transactions), it becomes very difficult to know when each
write completes.  I don't think that batching is necessary, 
as writes will accummulate in the socket buffer, so this
should be a better approach.

BUG=none
TEST=flip_session_unittest.cc

Review URL: http://codereview.chromium.org/342088

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@30898 0039d316-1c4b-4281-b951-d872f2087c98