Use TabSwitchPaintDuration in tab switching test.
Modified the tab switching test to use the new tab
switch paint duration histogram in place of the whiteout
duration histogram.  As a measure, whiteout duration
would not carry a value if we kept the backing store
around for a tab, and we would end up with a result of
zero instead of a useful result.  Local results with the
new measure are consistent with the number of operations
we perform on the tabs.

The tab switching test now fails if it cannot locate
the correct histogram.  In that case, it will output
the contents of the log file for remote debugging.

BUG=4104
TEST=count of tab switch paint duration histogram
results in tab switching test equals number of tabs
used in test

Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=32196

Review URL: http://codereview.chromium.org/385121

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@32543 0039d316-1c4b-4281-b951-d872f2087c98