linux: obey move.rects_valid in WebPluginGeometry
When move.rects_valid is false, we shouldn't look at the rects included
in the structure.

I am pretty sure the previous code was incorrect, but I am not sure why this
started mattering.  I know that plugins stopped displaying, but when I found
the regression window and reverted the change that first caused it, the problem
still happened.  It seems probable that another change that occurred while this
was regressed broke it further.

BUG=32965

Review URL: http://codereview.chromium.org/549149

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@37033 0039d316-1c4b-4281-b951-d872f2087c98