Attempt to fix remaining 3 tests for RSS.
I think the reason the negative feed parsing tests failed on the
bots is because the RSS extension has a generated ID that is
different from when you load it on a dev machine. So I now ask the
ExtensionService to provide the id during the run.

BUG=http://crbug.com/37109
TEST=Enabling test (although as flaky for now, until I get more
confidence that they are working).

Review URL: http://codereview.chromium.org/660424

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@40487 0039d316-1c4b-4281-b951-d872f2087c98