GTK: fix a tab dragging crasher.
The crash was a result of the early exit in EndDragImpl. When EndDragImpl is called due to a TAB_CONTENTS_DESTROYED notification and we exit early the DragController() doesn't get destroyed. We get a Drag() call afterwords, and call into a stale dragged_contents_. I fixed this by removing the early exit (instead checking dragged_tab_ before trying to clean up the drag, in case it hadn't started).

The change to EnsureDraggedTab() should not have any functional effect; it just makes the control flow clearer.

The new dragged_controller_ check should also not have any functional effect; it just seems safer to check it.

BUG=25326
TEST=manual (see bug for repro steps)
Review URL: http://codereview.chromium.org/668010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@40589 0039d316-1c4b-4281-b951-d872f2087c98