GTK: Use new macros to hide thunk declaration and cleanup callback declaration.
We have two conventions for declaring GTK+ callbacks in our code: We either
declare a bunch of static thunks in the header that redirect to the correct
object (and bloat up the header) OR we declare the callback methods static and
pass the object as the final parameter (and litter the implementation with
"context->"). The two styles are now about half/half...including mixing these
two styles in the same files. This madness must end!

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/661010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@41202 0039d316-1c4b-4281-b951-d872f2087c98