First cut at adding Extension Infobars to the experimental API.
Originally described here:
http://code.google.com/p/chromium/wiki/InfoBarExtensionAPI

The API is simple:

chrome.experimental.infoBar.show(
  {"htmlPath": "infobar.html"}, function(window) {
  // |window| is where the infobar was shown.
});

To close it, you can simply call window.close() from within the InfoBar (infoBar.hide is not provided).

The api is privileged (not available to content scripts directly).

The minimum height for the infobar is regular height for our infobars (32 pixels), maximum is twice that and the actual size is determined by the content (within these constraints mentioned).

See screenshot in bug. The icon on the far left is an extension icon with a dropdown menu showing the menu that shows when you right-click on the extension icon in the Omnibox/browser action container.

BUG=26463
TEST=None (for now).

Review URL: http://codereview.chromium.org/1049001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@41806 0039d316-1c4b-4281-b951-d872f2087c98